requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
pypdf